Always treat HTTP 401 as unauthorized even if supplied a different status code
This changes the semantics of the supplied statusCode option to be
treated as an alternative, rather than a replacement, for the standard
HTTP 401.

This helps in generic code where we can't be sure we can persuade the
server to output the alternate error code, but we want to succeed
regardless if we aren't running in a browser.

Amends: 3600ef55f98b544e36ac7837952eb2c9a45d12e4